refactor aggregator to verifier, fix clippy
