+++ pallets/randomness-beacon/src/verifier.rs
